Change dialog to accept array of txs
