+++ src/languageProvider/locales/en_US.json
@@ -156,12 +156,13 @@
   "str.cancel": "Cancel",
   "str.changePassphrase": "Change your passphrase",
   "str.close": "Close",
-  "str.confirm": "OK",
+  "str.confirm": "Confirm",
   "str.copy": "Copy",
   "str.createEvent": "Create Event",
   "str.creator": "Creator",
   "str.d": "d",
   "str.date": "Date",
+  "str.delete": "Delete",
   "str.description": "Description",
   "str.empty": " ",
   "str.emptyTxHistory": "You do not have any transactions right now.",
@@ -352,12 +353,14 @@
   "tx.approveBotTransfer": "Approve BOT Transfer",
   "tx.resetApproval": "Reset Approval",
   "txConfirm.title": "Please Confirm Your Transaction",
-  "txConfirm.txFeeMsg": "You are about to {txAction} with a maximum transaction fee of {txFee} QTUM. Any unused transaction fees will be refunded to you.",
+  "txConfirm.txFeeMsg": "You are about to {txAction} with a maximum transaction fee of {txFee} QTUM.",
   "txConfirm.txOne": "Confirmation for transaction 1/2. You will be required to confirm another transaction when this transaction is successful.",
   "txConfirm.txTwo": "Confirmation for transaction 2/2.",
+  "txConfirm.selectWalletProvider": "Select your wallet provider:",
+  "txConfirm.transactionX": "Transaction #{txNumber}",
+  "txConfirm.txFeeReturnMsg": "Gas fees are the maximum amount that you will spend for the transaction. Any unused transaction fees will be refunded to you.",
   "txConfirmMsg.bet": "bet on {option}",
   "txConfirmMsg.create": "create an event",
-  "txConfirm.selectWalletProvider": "Select your wallet provider:",
   "txConfirmMsg.send": "send to address {address}",
   "txConfirmMsg.set": "set the result as {option}",
   "txConfirmMsg.vote": "vote on {option}",
