+++ src/languageProvider/locales/ko-KR.json
@@ -156,7 +156,6 @@
   "str.cancel": "취소",
   "str.changePassphrase": "암호 변경",
   "str.close": "닫기",
-  "str.confirm": "승인",
   "str.copy": "복사",
   "str.createEvent": "이벤트 만들기",
   "str.creator": "생성자",
