+++ src/languageProvider/locales/zh-Hans.json
@@ -156,7 +156,6 @@
   "str.cancel": "取消",
   "str.changePassphrase": "更改您的密码",
   "str.close": "关闭",
-  "str.confirm": "确认",
   "str.copy": "复制",
   "str.createEvent": "创建事件",
   "str.creator": "创建者",
