+++ src/stores/models/Transaction.js
@@ -2,18 +2,16 @@ import _ from 'lodash';
 
 import { Token, TransactionType } from 'constants';
 import { gasToQtum, satoshiToDecimal } from '../../helpers/utility';
+
 const { APPROVE_CREATE_EVENT, APPROVE_SET_RESULT, APPROVE_VOTE, BET, VOTE, SET_RESULT, FINALIZE_RESULT, WITHDRAW, WITHDRAW_ESCROW } = TransactionType;
 
-
-/*
-* Model for Transactions.
-* Represents pending actions to contracts that are awaiting acceptance by the blockchain.
-* Transactions are currently local to each user's machine.
-*/
+/**
+ * Model for Transactions.
+ */
 export default class Transaction {
-  type = '' // One of: [CREATEEVENT, APPROVECREATEEVENT, BET, SETRESULT, APPROVESETRESULT, VOTE, APPROVEVOTE, FINALIZERESULT, WITHDRAW, WITHDRAWESCROW, TRANSFER]
+  type = '' // One of: TransactionType
   txid = '' // Transaction ID assigned by the blockchain
-  status = '' // One of: [PENDING, SUCCESS, FAIL]
+  status = '' // One of: TransactionStatus
   createdTime = '' // UNIX timestamp when Transaction was created
   blockNum = 0 // Block number when Transaction was recorded on blockchain
   blockTime = '' // Block timestamp for blockNum
@@ -26,26 +24,26 @@ export default class Transaction {
   oracleAddress = '' // Oracle contract address associated with Transaction
   name = '' // Name of the event
   optionIdx = 0 // Result index used for Transaction. eg. For a bet, this would be the result index the user bet on.
-  token = '' // Token type used for Transaction. QTUM for BET. BOT for VOTE.
   amount = '' // Amount of token used
+  token = '' // Token type used for Transaction. QTUM for BET. BOT for VOTE.
   topic // The Topic object associated with the Transaction
   version = 0 // Current version of the contract. To manage deprecations later.
-
-  // for invalid option
-  localizedInvalid = {};
+  localizedInvalid = {}; // for invalid option
 
   constructor(transaction) {
     Object.assign(this, transaction);
     this.gasLimit = Number(this.gasLimit);
     this.gasPrice = Number(this.gasPrice);
     this.fee = gasToQtum(this.gasUsed);
+    this.amount = this.token === Token.BOT ? satoshiToDecimal(this.amount) : this.amount;
+
     const { topic, type, optionIdx } = transaction;
     if (topic && [APPROVE_SET_RESULT, APPROVE_VOTE, BET, VOTE, SET_RESULT, FINALIZE_RESULT].includes(type)) {
       this.name = topic.options[optionIdx];
     } else if ([WITHDRAW, WITHDRAW_ESCROW].includes(type)) {
       this.name = type;
     }
-    this.amount = this.token === Token.BOT ? satoshiToDecimal(this.amount) : this.amount;
+
     this.localizedInvalid = {
       en: 'Invalid',
       zh: '无效',
